[Test] DeleteHistories 테스트 코드 작성 및 QuestionTest 변경
