+++ src/test/java/qna/domain/QuestionTest.java
@@ -113,36 +113,4 @@ public class QuestionTest {
 
         answers.forEach(answer -> assertThat(answer.isDeleted()).isTrue());
     }
-
-    @DisplayName("deleteHistories 생성")
-    @ParameterizedTest
-    @MethodSource(value = "provideQuestionWithSelfAnswer")
-    void makeDeleteHistories(Question question) throws CannotDeleteException {
-        assertThat(question.isDeleted()).isFalse();
-
-        question.delete(UserTest.JAVAJIGI);
-        assertThat(question.isDeleted()).isTrue();
-
-        List<DeleteHistory> deleteHistories = question.makeDeleteHistories();
-        assertThat(deleteHistories).hasSize(3);
-    }
-
-    @DisplayName("질문이 삭제되지 않아 deleteHistories 생성 실패")
-    @ParameterizedTest
-    @MethodSource(value = "provideQuestionWithSelfAnswer")
-    void makeDeleteHistories_fail(Question question) {
-        assertThat(question.isDeleted()).isFalse();
-        assertThatThrownBy(() -> question.makeDeleteHistories())
-                .isInstanceOf(CannotDeleteException.class)
-                .hasMessage(Question.CANNOT_MAKE_DELETE_HISTORIES_MESSAGE);
-    }
-
-    private static Stream<Arguments> provideQuestionWithSelfAnswer() {
-        Question question = new Question("title1", "contents1").writeBy(UserTest.JAVAJIGI);
-        question.addAnswer(new Answer(UserTest.JAVAJIGI, question, "Answers Contents1"));
-        question.addAnswer(new Answer(UserTest.JAVAJIGI, question, "Answers Contents2"));
-        return Stream.of(
-                Arguments.of(question, UserTest.JAVAJIGI)
-        );
-    }
 }
